+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/data/audio/OratureCueType.kt
@@ -1,0 +1,10 @@
+package org.wycliffeassociates.otter.common.data.audio
+
+enum class OratureCueType {
+    UNKNOWN,
+    CHUNK,
+    VERSE,
+    CHAPTER_TITLE,
+    BOOK_TITLE,
+    LICENSE
+}
