+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/audio/OtterAudioFile.kt
@@ -1,0 +1,180 @@
+package org.wycliffeassociates.otter.common.domain.audio
+
+import org.slf4j.LoggerFactory
+import org.wycliffeassociates.otter.common.audio.*
+import org.wycliffeassociates.otter.common.data.audio.*
+import java.io.File
+import java.util.regex.Pattern
+
+class OratureAudioFile : AudioFile {
+
+    val logger = LoggerFactory.getLogger(OratureAudioFile::class.java)
+
+    private val markers = OratureMarkers()
+
+    private fun initializeCues() {
+        markers.import(OratureCueParser.parse(this))
+        logger.info("Parsed ${getCuesFromMap(OratureCueType.VERSE)} verses")
+    }
+
+    constructor() : super() {
+        initializeCues()
+    }
+
+    constructor(file: File, metadata: AudioMetadata) : super(file, metadata) {
+        initializeCues()
+    }
+
+    constructor(file: File) : super(file) {
+        initializeCues()
+    }
+
+    constructor(
+        file: File,
+        channels: Int = DEFAULT_CHANNELS,
+        sampleRate: Int = DEFAULT_SAMPLE_RATE,
+        bitsPerSample: Int = DEFAULT_BITS_PER_SAMPLE
+    ) : super(file, channels, sampleRate, bitsPerSample) {
+        initializeCues()
+    }
+
+    constructor(
+        file: File,
+        channels: Int = DEFAULT_CHANNELS,
+        sampleRate: Int = DEFAULT_SAMPLE_RATE,
+        bitsPerSample: Int = DEFAULT_BITS_PER_SAMPLE,
+        metadata: AudioMetadata
+    ) : super(file, channels, sampleRate, bitsPerSample, metadata) {
+        initializeCues()
+    }
+
+    override fun getCues(): List<AudioCue> {
+        return markers.getCues()
+    }
+
+    fun getMarker(type: OratureCueType): List<AudioMarker> {
+        return getCuesFromMap(type)
+    }
+
+    fun addCues(cues: List<AudioCue>) {
+
+    }
+
+    fun addVerseMarker(marker: VerseMarker) {
+        markers.addMarker(OratureCueType.VERSE, marker)
+    }
+
+    fun addChunkMarker(marker: ChunkMarker) {
+        markers.addMarker(OratureCueType.CHUNK, marker)
+    }
+
+    fun importMetadata(metadata: AudioMetadata) {
+        markers.import(OratureCueParser.parse(metadata))
+    }
+
+    fun importCues(cues: List<AudioCue>) {
+        markers.import(OratureCueParser.parse(cues))
+    }
+
+    override fun addCue(location: Int, label: String) {
+        markers.import(OratureCueParser.parse(listOf(AudioCue(location, label))))
+    }
+
+    @Synchronized
+    fun addVerseMarker(location: Int, label: String) {
+        val regex = Pattern.compile("^(\\d+)(?:-(\\d+))?$")
+        val match = regex.matcher(label.trim())
+        if (match.matches()) {
+            val verseStart = match.group(1).toInt()
+            val verseEnd = match.group(2)?.toInt() ?: verseStart
+            val marker = VerseMarker(verseStart, verseEnd, location)
+            addVerseMarker(marker)
+        }
+    }
+
+    fun clearChunkMarkers() {
+        clearCuesFromMap(OratureCueType.CHUNK)
+    }
+
+    fun clearVerseMarkers() {
+        clearCuesFromMap(OratureCueType.VERSE)
+    }
+
+    override fun update() {
+        metadata.clearMarkers()
+        markers.getCues().forEach { metadata.addCue(it) }
+        super.update()
+    }
+
+    private fun getCuesFromMap(type: OratureCueType): List<AudioMarker> {
+        return markers.getMarkers(type)
+    }
+
+    private fun clearCuesFromMap(type: OratureCueType) {
+        markers.clearMarkersOfType(type)
+    }
+}
+
+class OratureMarkers {
+    private val cueMap: MutableMap<OratureCueType, MutableList<AudioMarker>> = mutableMapOf(
+        OratureCueType.CHUNK to mutableListOf(),
+        OratureCueType.VERSE to mutableListOf(),
+        OratureCueType.CHAPTER_TITLE to mutableListOf(),
+        OratureCueType.BOOK_TITLE to mutableListOf(),
+        OratureCueType.LICENSE to mutableListOf(),
+    )
+
+    fun getCues(): List<AudioCue> {
+        return cueMap.values.flatMap { it.map { it.toCue() } }
+    }
+
+    @Synchronized
+    fun getMarkers(type: OratureCueType): List<AudioMarker> {
+        if (!cueMap.containsKey(type)) cueMap[type] = mutableListOf()
+        return cueMap[type]!!
+    }
+
+    fun addMarkers(type: OratureCueType, markers: List<AudioMarker>) {
+        if (!cueMap.containsKey(type)) cueMap[type] = mutableListOf()
+        cueMap[type]!!.addAll(markers)
+    }
+
+    fun addMarker(type: OratureCueType, marker: AudioMarker) {
+        if (!cueMap.containsKey(type)) cueMap[type] = mutableListOf()
+        cueMap[type]!!.add(marker)
+    }
+
+    fun clearMarkersOfType(type: OratureCueType) {
+        if (!cueMap.containsKey(type)) {
+            cueMap[type] = mutableListOf()
+            return
+        } else {
+            cueMap[type]!!.clear()
+        }
+    }
+
+    private fun addEntry(entry: Map.Entry<OratureCueType, MutableList<AudioMarker>>) {
+        if (!cueMap.containsKey(entry.key)) cueMap[entry.key] = mutableListOf()
+        cueMap[entry.key]!!.addAll(entry.value)
+    }
+
+    /**
+     * Deep copies markers into a new instance of OratureMarkers
+     */
+    fun copy(): OratureMarkers {
+        val newCopy = OratureMarkers()
+        cueMap.forEach { newCopy.addEntry(it) }
+        return newCopy
+    }
+
+    /**
+     * Copies all markers from the provided markers to the internal map of markers
+     *
+     * @param markers the markers to copy from
+     */
+    fun import(markers: OratureMarkers) {
+        markers.cueMap.entries.forEach {
+            addEntry(it)
+        }
+    }
+}
