+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/chunking/ChunkAudioUseCase.kt
@@ -2,7 +2,7 @@ package org.wycliffeassociates.otter.common.domain.chunking
 
 import java.io.File
 import org.wycliffeassociates.otter.common.audio.AudioCue
-import org.wycliffeassociates.otter.common.domain.audio.decorators.OratureAudioFile
+import org.wycliffeassociates.otter.common.domain.audio.OratureAudioFile
 import org.wycliffeassociates.otter.common.domain.resourcecontainer.project.ProjectFilesAccessor
 import org.wycliffeassociates.otter.common.persistence.IDirectoryProvider
 import org.wycliffeassociates.resourcecontainer.ResourceContainer
