+++ common/src/test/kotlin/org/wycliffeassociates/otter/common/domain/chunking/ChunkAudioUseCaseTest.kt
@@ -12,7 +12,7 @@ import org.junit.rules.TemporaryFolder
 import org.wycliffeassociates.otter.common.*
 import org.wycliffeassociates.otter.common.audio.*
 import org.wycliffeassociates.otter.common.data.workbook.Translation
-import org.wycliffeassociates.otter.common.domain.audio.decorators.OratureAudioFile
+import org.wycliffeassociates.otter.common.domain.audio.OratureAudioFile
 import org.wycliffeassociates.otter.common.domain.resourcecontainer.project.ProjectFilesAccessor
 import org.wycliffeassociates.otter.common.persistence.IDirectoryProvider
 import org.wycliffeassociates.otter.common.persistence.repositories.WorkbookRepository
